// frontend/src/app/api/models/comment-entity.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CommentDecryptionKeyEntity } from '../models/comment-decryption-key-entity';
import { UserEntity } from '../models/user-entity';
export interface CommentEntity {
  author: UserEntity;
  content: string;
  createdAt: string;
  decryptionKeys: Array<CommentDecryptionKeyEntity>;
  id: number;
  iv: string;
}